requirements.txt
oracledb
matplotlib
PyQt5
python-dotenv
pandas
cryptography